Refactor/stdint

+++ include/dstring.h
@@ -1,6 +1,7 @@
 #ifndef DSTRING_H
 #define DSTRING_H
 
+#include <stdint.h>
 #include <stddef.h>
 
 typedef struct {
@@ -12,7 +13,7 @@ typedef struct {
 void * dstrinit(DString * str, char const * string, size_t const length);
 void dstrfree(DString * str);
 
-int dstrcmp(DString const * const str1, DString const * const str2);
+int8_t dstrcmp(DString const * const str1, DString const * const str2);
 DString * dstrcat(DString * dest, DString const * const src);
 DString * dstrcpy(DString * dest, DString const * const src);
 DString * dstrdup(DString const * const src);
